refactor(uniswap): rename files
